c325: extract LdaSweep family from api/client.ts to api/lda-sweep.ts (#441 P1 2.4)
api/client.ts 1059 LOC -> 1029 LOC (-30 LOC). New file: api/lda-sweep.ts (52 LOC).
Moved LdaSweepEntry + LdaSweep types and the ldaSweep runtime method.
client.ts re-exports types for back-compat; api.ldaSweep delegates to
ldaSweepApi.ldaSweep under the hood. typecheck clean, 44/44 tests pass.

This closes the audit-called-out trio (Routed + RateDistortion +
LdaSweep) of #441 P1 2.4. api/client.ts now down to 1029 LOC from the
original 1392 LOC — a -26% reduction across c261 (bandmask + eda + hidsag)
+ c323-c325 (this sweep).
